integration:: update package version to 1.0.0 and add Supabase dependencies
- Bump project version from 0.1.0 to 1.0.0 in package.json and package-lock.json.
- Introduce new dependencies for Supabase: `@supabase/auth-ui-react`, `@supabase/auth-ui-shared`, `@supabase/ssr`, and `@supabase/supabase-js`.
- Refactor page.tsx to utilize the new authentication hooks and streamline component structure.
