+++ components/landing/landing-page.tsx
@@ -1,0 +1,247 @@
+"use client";
+
+import { motion } from "framer-motion";
+import { Button } from "@/components/ui/button";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { ThemeToggle } from "@/components/theme-toggle";
+import {
+  BarChart3,
+  Shield,
+  TrendingUp,
+  Users,
+  Zap,
+  CheckCircle,
+  ArrowRight,
+  Wallet,
+  PieChart,
+  Target,
+} from "lucide-react";
+import AuthComponent from "@/components/auth/auth-component";
+
+interface LandingPageProps {
+  onStartApp: () => void;
+}
+
+const features = [
+  {
+    icon: <BarChart3 className="w-6 h-6" />,
+    title: "Control Total",
+    description:
+      "Rastrea todos tus gastos con categorías personalizables y análisis detallados.",
+  },
+  {
+    icon: <TrendingUp className="w-6 h-6" />,
+    title: "Análisis Inteligente",
+    description:
+      "Visualiza tus patrones de gasto y toma decisiones financieras informadas.",
+  },
+  {
+    icon: <Shield className="w-6 h-6" />,
+    title: "Datos Seguros",
+    description:
+      "Tu información financiera protegida con encriptación de nivel bancario.",
+  },
+  {
+    icon: <Zap className="w-6 h-6" />,
+    title: "Interfaz Rápida",
+    description:
+      "Agrega gastos en segundos con nuestra interfaz optimizada y moderna.",
+  },
+];
+
+const stats = [
+  { value: "100%", label: "Gratis" },
+  { value: "🔒", label: "Seguro" },
+  { value: "⚡", label: "Rápido" },
+  { value: "📱", label: "Responsive" },
+];
+
+export default function LandingPage({ onStartApp }: LandingPageProps) {
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-orange-50 via-white to-orange-50 dark:from-gray-900 dark:via-gray-800 dark:to-gray-900">
+      {/* Header */}
+      <header className="fixed top-0 w-full bg-white/80 dark:bg-gray-900/80 backdrop-blur-md border-b z-50">
+        <div className="container mx-auto px-4 h-16 flex items-center justify-between">
+          <motion.div
+            initial={{ opacity: 0, x: -20 }}
+            animate={{ opacity: 1, x: 0 }}
+            className="flex items-center space-x-3"
+          >
+            <div className="w-8 h-8 bg-orange-500 rounded-lg flex items-center justify-center">
+              <BarChart3 className="w-5 h-5 text-white" />
+            </div>
+            <span className="text-xl font-bold">Cifra</span>
+          </motion.div>
+
+          <motion.div
+            initial={{ opacity: 0, x: 20 }}
+            animate={{ opacity: 1, x: 0 }}
+          >
+            <ThemeToggle />
+          </motion.div>
+        </div>
+      </header>
+
+      {/* Hero Section */}
+      <section className="pt-32 pb-20 px-4">
+        <div className="container mx-auto text-center">
+          <motion.div
+            initial={{ opacity: 0, y: 30 }}
+            animate={{ opacity: 1, y: 0 }}
+            transition={{ duration: 0.8 }}
+            className="max-w-4xl mx-auto"
+          >
+            <h1 className="text-5xl md:text-6xl font-bold mb-6 bg-gradient-to-r from-orange-500 to-orange-600 bg-clip-text text-transparent">
+              Tu plata, bajo control
+            </h1>
+            <p className="text-xl md:text-2xl text-muted-foreground mb-8 max-w-2xl mx-auto">
+              La forma más inteligente de gestionar tus finanzas personales.
+              Controla gastos, visualiza patrones y alcanza tus metas
+              financieras.
+            </p>
+
+            <motion.div
+              initial={{ opacity: 0, y: 20 }}
+              animate={{ opacity: 1, y: 0 }}
+              transition={{ delay: 0.3 }}
+              className="flex flex-col sm:flex-row gap-4 justify-center items-center"
+            >
+              <AuthComponent onAuthSuccess={onStartApp} />
+              <div className="flex items-center text-sm text-muted-foreground">
+                <CheckCircle className="w-4 h-4 mr-2 text-green-500" />
+                Completamente gratis
+              </div>
+            </motion.div>
+          </motion.div>
+        </div>
+      </section>
+
+      {/* Stats Section */}
+      <section className="py-16 px-4">
+        <div className="container mx-auto">
+          <motion.div
+            initial={{ opacity: 0, y: 20 }}
+            whileInView={{ opacity: 1, y: 0 }}
+            viewport={{ once: true }}
+            className="grid grid-cols-2 md:grid-cols-4 gap-8 max-w-2xl mx-auto"
+          >
+            {stats.map((stat, index) => (
+              <motion.div
+                key={index}
+                initial={{ opacity: 0, scale: 0.8 }}
+                whileInView={{ opacity: 1, scale: 1 }}
+                viewport={{ once: true }}
+                transition={{ delay: index * 0.1 }}
+                className="text-center"
+              >
+                <div className="text-3xl font-bold text-orange-500 mb-2">
+                  {stat.value}
+                </div>
+                <div className="text-sm text-muted-foreground">
+                  {stat.label}
+                </div>
+              </motion.div>
+            ))}
+          </motion.div>
+        </div>
+      </section>
+
+      {/* Features Section */}
+      <section className="py-20 px-4">
+        <div className="container mx-auto">
+          <motion.div
+            initial={{ opacity: 0, y: 20 }}
+            whileInView={{ opacity: 1, y: 0 }}
+            viewport={{ once: true }}
+            className="text-center mb-16"
+          >
+            <h2 className="text-3xl md:text-4xl font-bold mb-4">
+              Todo lo que necesitas para controlar tus finanzas
+            </h2>
+            <p className="text-lg text-muted-foreground max-w-2xl mx-auto">
+              Herramientas profesionales diseñadas para simplificar tu gestión
+              financiera diaria.
+            </p>
+          </motion.div>
+
+          <div className="grid md:grid-cols-2 lg:grid-cols-4 gap-6">
+            {features.map((feature, index) => (
+              <motion.div
+                key={index}
+                initial={{ opacity: 0, y: 20 }}
+                whileInView={{ opacity: 1, y: 0 }}
+                viewport={{ once: true }}
+                transition={{ delay: index * 0.1 }}
+                whileHover={{ y: -5 }}
+              >
+                <Card className="h-full border-0 shadow-lg hover:shadow-xl transition-all duration-300">
+                  <CardHeader>
+                    <div className="w-12 h-12 bg-orange-100 dark:bg-orange-900/20 rounded-lg flex items-center justify-center text-orange-600 dark:text-orange-400 mb-4">
+                      {feature.icon}
+                    </div>
+                    <CardTitle className="text-lg">{feature.title}</CardTitle>
+                  </CardHeader>
+                  <CardContent>
+                    <CardDescription className="text-base leading-relaxed">
+                      {feature.description}
+                    </CardDescription>
+                  </CardContent>
+                </Card>
+              </motion.div>
+            ))}
+          </div>
+        </div>
+      </section>
+
+      {/* CTA Section */}
+      <section className="py-20 px-4">
+        <div className="container mx-auto">
+          <motion.div
+            initial={{ opacity: 0, y: 20 }}
+            whileInView={{ opacity: 1, y: 0 }}
+            viewport={{ once: true }}
+            className="bg-gradient-to-r from-orange-500 to-orange-600 rounded-2xl p-8 md:p-12 text-white text-center"
+          >
+            <h3 className="text-3xl md:text-4xl font-bold mb-4">
+              ¿Listo para tomar control de tus finanzas?
+            </h3>
+            <p className="text-xl mb-8 opacity-90 max-w-2xl mx-auto">
+              Únete a miles de usuarios que ya han mejorado su salud financiera
+              con Cifra.
+            </p>
+            <Button
+              size="lg"
+              variant="secondary"
+              onClick={onStartApp}
+              className="bg-white text-orange-600 hover:bg-gray-100 font-semibold px-8 py-3"
+            >
+              Empezar Ahora
+              <ArrowRight className="w-5 h-5 ml-2" />
+            </Button>
+          </motion.div>
+        </div>
+      </section>
+
+      {/* Footer */}
+      <footer className="py-8 px-4 border-t bg-white/50 dark:bg-gray-900/50">
+        <div className="container mx-auto text-center">
+          <div className="flex items-center justify-center space-x-2 mb-4">
+            <div className="w-6 h-6 bg-orange-500 rounded flex items-center justify-center">
+              <BarChart3 className="w-4 h-4 text-white" />
+            </div>
+            <span className="font-semibold">Cifra</span>
+          </div>
+          <p className="text-sm text-muted-foreground">
+            © 2024 Cifra. Tu plata, bajo control.
+          </p>
+        </div>
+      </footer>
+    </div>
+  );
+}
